requirement.txt
langchain-openai
langchain-community
langchain-ollama
python-dotenv
langchain_core
langchain
streamlit
ipykernel
langchain-ollama
langchain-groq
chromadb
faiss-cpu
langchain_chroma
ollama
pandas
bs4
sentence-transformers
langchain_huggingface
duckdb
arxiv
langchain-text-splitters
pypdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators
youtube_transcript_api
unstructured
pytube
numexpr
duckduckgo-search
U ddgs


